Rewrite to using 16x16 maps
This requires a different method for determining out-of-bounds accesses;
instead of having a 1-tile border, we have a separate "validmap" that
stores 1-bit per direction, which specifies whether the tile is valid in
that direction. It matches the other similar bitfield:

  2 4 1
  6   7
  3 5 0

All operations that move around the board need to first check this map
before moving.