Optimize calcdist; flip VALID_* and MASK_*
I realized that calcdist() was using flags_bin directly to determine if
the neighbors were walkable, but that information is already cached in
sigmap.

However, VALID_* and MASK_* were actually reversed, since I was using
the bits as set from the opposite direction. After flipping them, I was
able to mask the signature and valid mask to quickly determine a valid
direction to move.

The other major optimization is by handling all tiles of a given
distance before moving onto the next. This means we don't have to store
the distance, since we always handle distance 4 before 5 before 6, etc.
This is easy to do by adding a second tail; the newtail adds candidates
of the next distance, and the oldtail tracks the current distance.