Modify room generation routine to allow gate rooms
The previous room generation required keeping track of rectangular
regions where a room can be placed. This falls over when there are
existing tiles already (as with a gate room). It would have been
possible to have pre-defined regions in this case, but I decided to use
a different algorithm instead. I've also tried to update some of the
hard-coded values to use defines instead.

The gate rooms also used a special wall tile for a fixed wall that
cannot be carved; this requires special handling in some routines.