Add pickups
They share a lot of similarities with mobs, but don't have AI, hp, etc.
This handles the basic drawing/animation/movement of pickups, but
nothing else yet.

The inline assembler here is an alternate way to determine if a relative
movement from a known good position is valid. It relies on the H and C
flags being set when wrapping in the X and Y dimensions respectively.
Seems like there may be a better way to do it, but for now this works.
And for the most part, a pickup will be able to drop in the same
location it came from (the chest, vase, etc.)