Implement reaper and messages
The only messages are: "wurstlord awakens" and "one monster holds the
key", both of which require 18 sprites. The first 18 sprites are always
allocated for a message. The next 8 sprites are always allocated for
floats. That leaves only 14 sprites for mob animation, but that seems to
be plenty for now.

The reaper appears after 100 steps (not including bumps/attacks).

Fixed a couple of bugs too:
* Picking up a "slap" would crash the game, since the
  pick_type_name_start array didn't have an entry for N+1 (so the length
  could be calculated).
* The number of keys wasn't updated in the HUD when you open the gate
* If the last mob had a key, then the key_mob wasn't cleared to 0, so
  it was possible to spawn another key mob
* num_dead_mobs was never being initialized