Add blinding
Blinding reduces the player's sight distance to 1, and resets the "fog"
map. A player is always blinded if they are attacked by a scorpion, or
w/ a 50% chance when they attack a scorpion. Blindness recovers after 30
frames.

The updated sight() function now works when blinded is that it continues
to expand to a distance of 4 for mobs, but doesn't unfog anything past
the first 4 surrounding tiles.

Also:
* Rename sightmap to mobsightmap, since it is always a distance of 4
  (but the player may have a sight distance of only 1)
* Rename *_TIME -> *_FRAMES or *_MOVES as appropriate
* Rename num_floats -> next_foat, since it starts at BASE_FLOAT
* Fix a bug in update_floats_and_msg() which wasn't properly removing
  old floats