Exploding bombs
Bombs are "objects" that have drops. When hit, they damage the
surrounding 8 tiles and spawn some sprites.

This change adds a new list of "spr" which are used for this (and will
be used for ability effects too).

Each spr's x, y, dx, and dy are stored as 8.8 fixed point.

Also:
* Add flags.c and map.c (they are no longer generated, but I forgot to
  add them because they were in .gitignore)
* Add util.s for assembly functions (moved is_valid out since it wasn't
  really inline asm)
* Always reset animation speed (after move/bump etc.) Otherwise it is
  possible for an active mob that is fogged to call mobwalk(), set a
  fast animation speed, then continue to animate quickly when unfogged.