Pass turns more quickly
The previous method would wait a frame to process mobs/weeds, but now we
always check mobs/weeds at the end of the player's movement. If the
mobs/weeds actually take a turn, then it will animate them as before.

This makes the player movement feel less stiff. We could probably save
another frame by processing mob movement during animation, but that will
be a bit trickier since we'd have to save those actions for later.