Completely overhaul animation system
This change simplifies the tile update logic, which was especially
complicated when unfogging tiles or when standing on top of an animating
tile.

Use dirty tiles instead; every time we want to update a tile we mark it
in the dirtymap and the dirty tile list. If the tile is already in the
dirtymap, we skip it.

Animated tiles are also added to the dirty tile list, and they are
updated just as any other tile would be; they just change tmap[] to
update the tile.

When dirty tiles are processed, they are checked to see whether they
should display a mob, pickup, or a tile (in that order). A mob or pickup
will not be drawn as a tile if they are currently being drawn as a
sprite (i.e. mob_move_timer[] or pick_move_timer[] are non-zero).

This means that the do_animate() function needs to change a bit --
previously do_ai() would run after the player moved, but do_animate()
would not be called a second time. This was leaving a frame where mobs
wouldn't be drawn at all since mobwalk() would be called on them, but
the sprite wouldn't be drawn yet.

Now do_animate(), pass_turn() and do_ai() are called in a loop (instead
of recursion), and do_animate() will be called after do_ai() if any mobs
moved.

Some renamings:
  do_animate() -> animate()
  do_ai() -> ai_run_tasks()
  do_mob_ai() -> ai_run_mob_task()

Since tile updating only happens for the play screen, all other tile
updating (e.g. inventory screen) have been switched over to calls to
set_vram_byte() or vram_copy().