First pass on sfx
Some sfx don't really work via the automatic conversion, need to be
tweaked. It's hard to get the noise channel to sound good automatically.

Some bugs in transitioning from sfx -> music; need to restore
instrument, perhaps? Probably need a sfx priority system too, since we
can only play one sfx at a time.

Need to move the loud part of the second part of the song to channel 1;
right now it's on channel 2 so it is cut out when sfx plays.