Better randomization of levels
We used to set the random number seed when the player pressed the start
button on the title screen. The high byte was set from the `sys_time`
variable (i.e. a frame count) and the low byte was set from the DIV
register.

The problem with this solution is that it isn't very random, since we're
waiting for vblank, it happens that the DIV register is always the same
value for a given frame. Another problem is that you could just hold
down the start button to get a completely consistent floor 0 and floor
1.

To fix these issues, we first prevent the player from holding the start
button by making sure it isn't pressed first. To generate a better
random seed, we try to mix in more entropy on floor 0. Whenever the
player presses a button, we mix in the joypad state and the current DIV
register into the seed. This seems to give a much less consistent layout
for floor 1.