Add repeat rate, and buffered input
The pico-8 version uses btnp() for various functions, which will return
true on the 1st frame it is pressed, then 15 frames later, and then
every 4 frames after that. That is measured based on a 30fps game, so to
mimic this behavior we repeat at 30 and 8 frames instead.

The repeat rate works well to prevent accidental movement, but it can
feel like you're inputs are dropped, since it takes ~8 frames to move
between tiles (longer if enemies are taking a turn). The pico-8 version
fixes this by buffering one action.