Fix bug where mob AI is skipped
Some attacks (slap and bolt) call hitpos(), which does damage and stuns
a mob. Before this change, if the mob died as a result of the attack,
then the mob that swapped into its slot would end up being stunned. This
often wouldn't be noticeable since that mob may not even be attacking
the player.

The fix is to always stun the mob before doing damage.