Simplify main turn loop
It took me a while to remember how this all works. But basically, the
main loop does this:

* get player movement
* wait for animations to finish
* get mob movements
* wait for animations to finish
* get plant movements
* wait for animations to finish

But it also will skip ahead if there are no currently running animations
for the mobs/plants.

The pico8 version uses callback functions to do this, but it's easy
enough to just use a loop. Originally I wrote it to use `doai` and
`dopassturn` variables, but that makes things more confusing. This
commit changes that so each function returns a value to determine
whether to continue with the state machine or not.