Optimize sight and is_valid
The previous `sight()` algorithm was implemented as a BFS, starting from
the player position and shooting rays out according to the data in
`sightsig`. This is simpler in some ways because we can take a given
position, move it one step away (in one of 8 directions), and add it to
the queue to process later. But it's also more expensive than necessary;
each of the 8 directions must be tested even though most of them are
never used.

Instead, if we follow a given ray as long as it doesn't hit a wall, then
the next, etc. we can do a lot better. Two arrays are used for this:
`sightdiff` and `sightskip`. `sightdiff` gives a list of the relative
offsets of a ray. They're all packed together, however. So when a ray
hits a wall, `sightskip` tells how many of the next elements of
`sightdiff` to skip over (these would have been the continuation of the
previous ray).

For this to work, we need to be able to efficiently determine whether
`pos + diff` is valid; i.e. in bounds. I wrote a function to do this
before, `is_valid`, but it was completely broken. I've since rewritten
it and it seems to work now.

It ends up using a somewhat clever technique, but I'm not sure whether
it's actually faster than the naive one. Basically we want to know
whether `pos + diff` has overflowed in either the X or Y direction.
Since pos is stored as `0xYX`, we can attempt to detect overflow in both
directions at the same time. We use a formula from Hacker's Delight:

```
sum = a + b
overflow = (sum ^ a) & (sum ^ b) & 8
```

Where the overflow ends up in the sign bit (bit 3 in our case). However
this formula assumes that `a` and `b` are both signed. In our case `pos`
has range [0, 15] and `diff` has range [-8, 7]. To make the math work
out in this case we need to modify it to:

```
pos' = pos ^ 8
sum = pos' + diff
overflow = (sum ^ pos') & (sum ^ diff) & 8
```

To calculate overflow for both X and Y in parallel, we can do the
following:

```
pos' = pos ^ 0x88
sum = pos' + diff
overflow = (sum ^ pos') & (sum ^ diff) & 0x88
```

However this doesn't take into account the fact that `pos' + diff` may
overflow from X into Y. Fortunately we can detect this in hardware as
the half-carry flag! Unfortunately, it's a huge pain to do so.

The half-carry flag is only read during the DAA instruction. If the
half-carry flag is set, then '6' will be added to the `A` register. But
`6` will _also_ be added if the sum in the low nibble is greater than
`9`! We don't want that. But it turns out that this doesn't happen if
the negative flag (`N`) is set.

So we want to make sure the `N` flag is set before we execute the `DAA`
instruction. Unfortunately, many instructions set this flag, so the only
way we can be sure that it's set is by subtracting instead of adding. So
we can change our formula to:

```
pos' = pos ^ 0x88
sum = pos' - (-diff)
overflow = (sum ^ pos') & (sum ^ diff) & 0x88
```

To calculate `-diff` for X and Y in parallel, we use start with the
basic formula:

```
-a = ~a + 1
```

So for both nibbles it would be:

```
-diff = ~diff + 0x11
```

Unfortunately, the sum in the low nibble can overflow. We can work
around this by adding `0x10` then swapping the nibbles, then adding
`0x10` again and swapping the nibbles again.

Now if the `N` and `H` flags are set, then `DAA` will add `0xA` to the
`A` register. So we set it to `0` carefully (making sure not to touch
any flags), and then we can left-shit `A` and mask with `0x10` to
extract the half-carry. If there was a half-carry then we subtracted too
much, so we add back in `0x10`. Now we can use this corrected sum for
the rest of the calculation. So the final formula:

```
pos' = pos ^ 0x88
sum = pos' - (-diff)
if half-carry:
  sum += 0x10
overflow = (sum ^ pos') & (sum ^ diff) & 0x88
```